refactor(hero): drop chain CSS rules in favor of composition
Instead having several classes in each of the DOM elements, in the project we will favor the
creation of a single class that gather several styles and ship them together. The usage for
several classes will be reduced to perform layout changes (container, columns...) and adding
modifiers for certain cases of a class over few properties.