feat(informative): add demand content for informative section
This code, doesn't intentionally allow the user to change the structure of the content in the
section but force the design suggested structure. In order to accomplish more freedom we may want to
introduce "markdown" widget on the content of this section, such decision need to be further
discussed and taking into count that the user can break the design.