fix(hero): allow section to fit into screen in desktop devices
by adding a padding into "main" at 11d3ccea429560579613be271532e08125b94669 lead us to have a hero
section that doesn't fit into the screen. Instead, we will move the responsability to each page to
have the desire space. In order to have a uniform space we use a variable of "$header-height-md"
which is not available unless we include the styles for data dues from within styles index instead